python/pyproject.toml
[project]
name = "silent-sky"
version = "0.1.0"
description = "Space observatory POMDP environment"
requires-python = ">=3.8"
dependencies = [
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "pyzmq>=25.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

